#H5IS
Too much refactoring - too little testing. I copied a few lines from one of the other method in the class, but didn't test the changes properly. The "Set" method should use the indexer of the underlying dictionary so that we set the value (and thereby replace any existing value) opposed to try and the value even if there already is an existing value.