Removed the SearchUtils class
The method contained a few static methods for helping with sorting instances ISearchResult by a date. The implemention would try to parse a string value to a DateTime instance.

Generally speaking it can be expensive in terms of performance to parse many DateTime values / especially via a request where execution time matter. The implementions did only try a single date format via DateTime.TryParseExact, which is a lot better than DateTime.TryParse, as the latter tries a number of different formats which adds to the parsing time.

But as we via other parts of the package usually make sure to add date and time values to the index in a searchable/sortable format, we can sort by this instead, which means we don't have to parse the indexed string value back to a DateTime instance. Which means a win in performance.